# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basis Social API",
    "version": "1.0.0-alpha",
    "description": "Stable REST and world-server integration contract for BasisVR. ActivityPub endpoints are intentionally outside this client API contract."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
    "/api/v1/auth/register": {"post": {"tags":["Auth"],"operationId":"register","requestBody":{"$ref":"#/components/requestBodies/Register"},"responses":{"201":{"description":"Registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSession"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/auth/login": {"post": {"tags":["Auth"],"operationId":"login","requestBody":{"$ref":"#/components/requestBodies/Login"},"responses":{"200":{"description":"Authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSession"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/auth/refresh": {"post": {"tags":["Auth"],"operationId":"refreshSession","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["refreshToken"],"properties":{"refreshToken":{"type":"string"}}}}}},"responses":{"200":{"description":"Refreshed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSession"}}}},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/auth/logout": {"post": {"tags":["Auth"],"operationId":"logout","security":[{"bearerAuth":[]}],"responses":{"204":{"description":"Logged out"},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/me": {
      "get": {"tags":["Profiles"],"operationId":"getCurrentUser","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Current user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Profile"}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "delete": {"tags":["Profiles"],"operationId":"deleteCurrentAccount","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["password"],"properties":{"password":{"type":"string","minLength":1}}}}}},"responses":{"204":{"description":"Account content removed and identity anonymized"},"default":{"$ref":"#/components/responses/Error"}}}
    },
    "/api/v1/me/export": {"get":{"tags":["Profiles"],"operationId":"exportCurrentAccount","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Machine-readable account data export","headers":{"Content-Disposition":{"schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"object","required":["schemaVersion","generatedAt","account"],"properties":{"schemaVersion":{"type":"integer"},"generatedAt":{"type":"string","format":"date-time"},"account":{"type":"object"},"profile":{"type":"object"},"actor":{"type":"object"},"relationships":{"type":"array","items":{}},"worlds":{"type":"array","items":{}},"events":{"type":"array","items":{}},"invites":{"type":"array","items":{}},"groupMemberships":{"type":"array","items":{}},"notifications":{"type":"array","items":{}}}}}}},"default":{"$ref":"#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type":"http","scheme":"bearer","bearerFormat":"JWT"},
      "serviceToken": {"type":"apiKey","in":"header","name":"X-Basis-Service-Token"}
    },
...
    "responses": {
      "Error": {"description":"Error response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}
    },
    "requestBodies": {
      "Register": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["username", "email", "password"],
              "properties": {
                "username": {"type": "string"},
                "email": {"type": "string", "format": "email"},
                "password": {"type": "string", "format": "password"}
              }
            }
          }
        }
      },
      "Login": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["login", "password"],
              "properties": {
                "login": {"type": "string"},
                "password": {"type": "string", "format": "password"}
              }
            }
          }
        }
      }
    },
...
    "schemas": {
      "Error": {"type":"object","required":["error"],"properties":{"error":{"type":"object","required":["code","message"],"properties":{"code":{"type":"string"},"message":{"type":"string"}}}}},
      "AuthSession": {"type":"object","required":["accessToken","refreshToken"],"properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/Profile"}}},
      "Profile": {"type":"object","properties":{"id":{"type":"string","format":"uuid"},"actorId":{"type":"string","format":"uuid"},"username":{"type":"string"},"acct":{"type":"string"},"displayName":{"type":"string"},"bio":{"type":"string"},"avatarUrl":{"type":"string"},"bannerUrl":{"type":"string"},"statusText":{"type":"string"},"links":{"type":"array","items":{}},"relationship":{"type":"object"}}},
...
    }
  }
}